converted all files under directoryServices to ts
