+++ src/services/directoryServices/MediaDirectoryService.ts
@@ -1,41 +1,66 @@
-const { BadRequestError } = require("@errors/BadRequestError")
+import { BadRequestError } from "@errors/BadRequestError"
 
-const { isMediaPathValid } = require("@validators/validators")
+import { isMediaPathValid } from "@validators/validators"
+
+import GithubSessionData from "@root/classes/GithubSessionData"
+import UserWithSiteSessionData from "@root/classes/UserWithSiteSessionData"
+import { GitDirectoryItem } from "@root/types/gitfilesystem"
+import { MediaDirOutput, MediaFileOutput } from "@root/types/media"
+import RepoService from "@services/db/RepoService"
+import BaseDirectoryService from "@services/directoryServices/BaseDirectoryService"
 
 const PLACEHOLDER_FILE_NAME = ".keep"
 
 class MediaDirectoryService {
-  constructor({ baseDirectoryService, gitHubService }) {
+  private baseDirectoryService: BaseDirectoryService
+
+  private repoService: RepoService
+
+  constructor({
+    baseDirectoryService,
+    repoService,
+  }: {
+    baseDirectoryService: BaseDirectoryService
+    repoService: RepoService
+  }) {
     this.baseDirectoryService = baseDirectoryService
-    this.gitHubService = gitHubService
+    this.repoService = repoService
   }
 
-  /**
-   * Lists files in directory. Returns empty array if directory does not exist
-   * - useful for base media directories which do not have placeholder files
-   */
-  async listWithDefault(sessionData, { directoryName }) {
-    let files = []
+  async listWithDefault(
+    sessionData: UserWithSiteSessionData,
+    { directoryName }: { directoryName: string }
+  ): Promise<GitDirectoryItem[]> {
+    let files: GitDirectoryItem[] = []
     try {
       const retrievedFiles = await this.baseDirectoryService.list(sessionData, {
         directoryName,
       })
       files = retrievedFiles
-    } catch (error) {
-      // return an empty list if directory does not exist
+    } catch (error: any) {
       if (error.status !== 404) throw error
     }
     return files
   }
 
   async listMediaDirectoryContent(
-    sessionData,
-    { directoryName, page, limit, search }
-  ) {
-    if (!isMediaPathValid({ path: directoryName }))
+    sessionData: UserWithSiteSessionData,
+    {
+      directoryName,
+      page,
+      limit,
+      search,
+    }: { directoryName: string; page: number; limit: number; search: string }
+  ): Promise<{
+    directories: MediaDirOutput[]
+    files: Pick<MediaFileOutput, "name">[]
+    total: number
+  }> {
+    if (!isMediaPathValid({ path: directoryName })) {
       throw new BadRequestError("Invalid media folder name")
+    }
 
-    return this.gitHubService.readMediaDirectory(
+    return this.repoService.readMediaDirectory(
       sessionData,
       directoryName,
       page,
@@ -45,24 +70,25 @@ class MediaDirectoryService {
   }
 
   async createMediaDirectory(
-    sessionData,
-    githubSessionData,
-    { directoryName, objArray }
-  ) {
-    if (!isMediaPathValid({ path: directoryName }))
+    sessionData: UserWithSiteSessionData,
+    githubSessionData: GithubSessionData,
+    { directoryName, objArray }: { directoryName: string; objArray: any[] }
+  ): Promise<{ newDirectoryName: string }> {
+    if (!isMediaPathValid({ path: directoryName })) {
       throw new BadRequestError(
         "Special characters not allowed in media folder name"
       )
+    }
+
     if (directoryName === "images" || directoryName === "files") {
       throw new BadRequestError("Cannot create root media directory")
     }
 
     if (objArray && objArray.length !== 0) {
-      // We can't perform these operations concurrently because of conflict issues
-      /* eslint-disable no-await-in-loop, no-restricted-syntax */
       const pathTokens = directoryName.split("/")
       const oldDirectoryName = pathTokens.slice(0, -1).join("/")
       const targetFiles = objArray.map((file) => file.name)
+
       await this.baseDirectoryService.moveFiles(
         sessionData,
         githubSessionData,
@@ -75,27 +101,29 @@ class MediaDirectoryService {
       )
     }
 
-    // We do this step later because the git tree operation overrides it otherwise
-    await this.gitHubService.create(sessionData, {
+    await this.repoService.create(sessionData, {
       content: "",
       fileName: PLACEHOLDER_FILE_NAME,
       directoryName,
     })
 
-    return {
-      newDirectoryName: directoryName,
-    }
+    return { newDirectoryName: directoryName }
   }
 
   async renameMediaDirectory(
-    sessionData,
-    githubSessionData,
-    { directoryName, newDirectoryName }
-  ) {
-    if (!isMediaPathValid({ path: newDirectoryName }))
+    sessionData: UserWithSiteSessionData,
+    githubSessionData: GithubSessionData,
+    {
+      directoryName,
+      newDirectoryName,
+    }: { directoryName: string; newDirectoryName: string }
+  ): Promise<void> {
+    if (!isMediaPathValid({ path: newDirectoryName })) {
       throw new BadRequestError(
         "Special characters not allowed in media folder name"
       )
+    }
+
     await this.baseDirectoryService.rename(sessionData, githubSessionData, {
       oldDirectoryName: directoryName,
       newDirectoryName,
@@ -104,12 +132,14 @@ class MediaDirectoryService {
   }
 
   async deleteMediaDirectory(
-    sessionData,
-    githubSessionData,
-    { directoryName }
-  ) {
-    if (!isMediaPathValid({ path: directoryName }))
+    sessionData: UserWithSiteSessionData,
+    githubSessionData: GithubSessionData,
+    { directoryName }: { directoryName: string }
+  ): Promise<void> {
+    if (!isMediaPathValid({ path: directoryName })) {
       throw new BadRequestError("Invalid media folder name")
+    }
+
     await this.baseDirectoryService.delete(sessionData, githubSessionData, {
       directoryName,
       message: `Deleting media folder ${directoryName}`,
@@ -117,17 +147,23 @@ class MediaDirectoryService {
   }
 
   async moveMediaFiles(
-    sessionData,
-    githubSessionData,
-    { directoryName, targetDirectoryName, objArray }
-  ) {
+    sessionData: UserWithSiteSessionData,
+    githubSessionData: GithubSessionData,
+    {
+      directoryName,
+      targetDirectoryName,
+      objArray,
+    }: { directoryName: string; targetDirectoryName: string; objArray: any[] }
+  ): Promise<void> {
     if (
       !isMediaPathValid({ path: directoryName }) ||
       !isMediaPathValid({ path: targetDirectoryName })
-    )
+    ) {
       throw new BadRequestError(
         "Special characters not allowed in media folder name"
       )
+    }
+
     const targetFiles = objArray.map((item) => item.name)
 
     await this.baseDirectoryService.moveFiles(sessionData, githubSessionData, {
