+++ src/services/directoryServices/ResourceRoomDirectoryService.ts
@@ -1,37 +1,77 @@
-const { BadRequestError } = require("@errors/BadRequestError")
-const { ConflictError } = require("@errors/ConflictError")
+import { BadRequestError } from "@errors/BadRequestError"
 
-const {
+import {
   convertDataToMarkdown,
   retrieveDataFromMarkdown,
-} = require("@utils/markdown-utils")
-const { slugifyCollectionName } = require("@utils/utils")
+} from "@utils/markdown-utils"
+import { slugifyCollectionName } from "@utils/utils"
+
+import GithubSessionData from "@root/classes/GithubSessionData"
+import UserWithSiteSessionData from "@root/classes/UserWithSiteSessionData"
+import { ConflictError } from "@root/errors/ConflictError"
+import { GitDirectoryItem } from "@root/types/gitfilesystem"
+import GitHubService from "@services/db/GitHubService"
+import BaseDirectoryService from "@services/directoryServices/BaseDirectoryService"
+import { ConfigYmlService } from "@services/fileServices/YmlFileServices/ConfigYmlService"
 
 const INDEX_FILE_NAME = "index.html"
 
+interface ConfigContent {
+  "facebook-pixel": string
+  resources_name?: string
+}
+
 class ResourceRoomDirectoryService {
-  constructor({ baseDirectoryService, configYmlService, gitHubService }) {
+  private baseDirectoryService: BaseDirectoryService
+
+  private configYmlService: ConfigYmlService
+
+  private gitHubService: GitHubService
+
+  constructor({
+    baseDirectoryService,
+    configYmlService,
+    gitHubService,
+  }: {
+    baseDirectoryService: BaseDirectoryService
+    configYmlService: ConfigYmlService
+    gitHubService: GitHubService
+  }) {
     this.baseDirectoryService = baseDirectoryService
     this.configYmlService = configYmlService
     this.gitHubService = gitHubService
   }
 
-  async listAllResourceCategories(sessionData, { resourceRoomName }) {
+  async listAllResourceCategories(
+    sessionData: UserWithSiteSessionData,
+    { resourceRoomName }: { resourceRoomName: string }
+  ): Promise<GitDirectoryItem[]> {
     const filesOrDirs = await this.baseDirectoryService.list(sessionData, {
-      directoryName: `${resourceRoomName}`,
+      directoryName: resourceRoomName,
     })
-    return filesOrDirs.reduce((acc, curr) => {
-      if (curr.type === "dir")
-        acc.push({
-          name: curr.name,
-          type: "dir",
-        })
-      return acc
-    }, [])
+    return filesOrDirs.reduce(
+      (acc: GitDirectoryItem[], curr: GitDirectoryItem) => {
+        if (curr.type === "dir") {
+          acc.push({
+            name: curr.name,
+            type: "dir",
+            path: "",
+            size: 0,
+            addedTime: 0,
+          })
+        }
+        return acc
+      },
+      []
+    )
   }
 
-  async getResourceRoomDirectoryName(sessionData) {
-    const config = await this.configYmlService.read(sessionData)
+  async getResourceRoomDirectoryName(
+    sessionData: UserWithSiteSessionData
+  ): Promise<{ resourceRoomName: string | null }> {
+    const config: { content: ConfigContent } = await this.configYmlService.read(
+      sessionData
+    )
     return {
       resourceRoomName: config.content.resources_name
         ? config.content.resources_name
@@ -39,20 +79,23 @@ class ResourceRoomDirectoryService {
     }
   }
 
-  async createResourceRoomDirectory(sessionData, { resourceRoomName }) {
+  async createResourceRoomDirectory(
+    sessionData: UserWithSiteSessionData,
+    { resourceRoomName }: { resourceRoomName: string }
+  ): Promise<{ newDirectoryName: string }> {
     if (/[^a-zA-Z0-9- ]/g.test(resourceRoomName)) {
-      // Contains non-allowed characters
       throw new BadRequestError(
         "Special characters not allowed in resource room name"
       )
     }
     const slugifiedResourceRoomName = slugifyCollectionName(resourceRoomName)
-    const { content: configContent, sha } = await this.configYmlService.read(
+    const { content: configContent, sha } = (await this.configYmlService.read(
       sessionData
-    )
-    // If resource room already exists, throw error
-    if ("resources_name" in configContent && configContent.resources_name)
+    )) as { content: ConfigContent; sha: string }
+
+    if ("resources_name" in configContent && configContent.resources_name) {
       throw new ConflictError("Resource room already exists")
+    }
     configContent.resources_name = slugifiedResourceRoomName
     await this.configYmlService.update(sessionData, {
       fileContent: configContent,
@@ -67,6 +110,7 @@ class ResourceRoomDirectoryService {
       content: newContent,
       fileName: INDEX_FILE_NAME,
       directoryName: slugifiedResourceRoomName,
+      isMedia: false,
     })
     return {
       newDirectoryName: slugifiedResourceRoomName,
@@ -74,18 +118,19 @@ class ResourceRoomDirectoryService {
   }
 
   async renameResourceRoomDirectory(
-    sessionData,
-    githubSessionData,
-    { resourceRoomName, newDirectoryName }
-  ) {
+    sessionData: UserWithSiteSessionData,
+    githubSessionData: GithubSessionData,
+    {
+      resourceRoomName,
+      newDirectoryName,
+    }: { resourceRoomName: string; newDirectoryName: string }
+  ): Promise<{ newDirectoryName: string }> {
     if (/[^a-zA-Z0-9- ]/g.test(newDirectoryName)) {
-      // Contains non-allowed characters
       throw new BadRequestError(
         "Special characters not allowed in resource category name"
       )
     }
     const slugifiedNewResourceRoomName = slugifyCollectionName(newDirectoryName)
-
     const { content: rawContent, sha } = await this.gitHubService.read(
       sessionData,
       {
@@ -96,7 +141,6 @@ class ResourceRoomDirectoryService {
     const { frontMatter, pageContent } = retrieveDataFromMarkdown(rawContent)
     frontMatter.title = newDirectoryName
     const newContent = convertDataToMarkdown(frontMatter, pageContent)
-
     await this.baseDirectoryService.rename(sessionData, githubSessionData, {
       oldDirectoryName: resourceRoomName,
       newDirectoryName: slugifiedNewResourceRoomName,
@@ -108,11 +152,13 @@ class ResourceRoomDirectoryService {
       fileName: INDEX_FILE_NAME,
       directoryName: slugifiedNewResourceRoomName,
     })
-
     const {
       content: configContent,
       sha: configSha,
-    } = await this.configYmlService.read(sessionData)
+    } = (await this.configYmlService.read(sessionData)) as {
+      content: ConfigContent
+      sha: string
+    }
     configContent.resources_name = slugifiedNewResourceRoomName
     await this.configYmlService.update(sessionData, {
       fileContent: configContent,
@@ -124,18 +170,17 @@ class ResourceRoomDirectoryService {
   }
 
   async deleteResourceRoomDirectory(
-    sessionData,
-    githubSessionData,
-    { resourceRoomName }
-  ) {
+    sessionData: UserWithSiteSessionData,
+    githubSessionData: GithubSessionData,
+    { resourceRoomName }: { resourceRoomName: string }
+  ): Promise<void> {
     await this.baseDirectoryService.delete(sessionData, githubSessionData, {
       directoryName: resourceRoomName,
       message: `Deleting resource room ${resourceRoomName}`,
     })
-
-    const { content: configContent, sha } = await this.configYmlService.read(
+    const { content: configContent, sha } = (await this.configYmlService.read(
       sessionData
-    )
+    )) as { content: ConfigContent; sha: string }
     delete configContent.resources_name
     await this.configYmlService.update(sessionData, {
       fileContent: configContent,
